builder and hp.config
