+++ core/utils/collect-hp-config.js
@@ -1,0 +1,19 @@
+const {mergeObjects} = require("./merge-objects");
+const path = require('path');
+
+function collectHPConfig(defaultOptions, options) {
+  let hbConfig = {};
+  
+  try {
+    hbConfig = require(path.join(process.cwd(), 'hp.config.js'));
+  } catch (e) {
+    if (e.code !== 'MODULE_NOT_FOUND') {
+      e.message = `hp.config.js file loading failed: ${e.message}`
+      throw new Error(e)
+    }
+  }
+  
+  return {...defaultOptions, ...mergeObjects(hbConfig, options)};
+}
+
+module.exports.collectHPConfig = collectHPConfig;
